+++ app/Http/Controllers/ProfileController.php
@@ -3,10 +3,13 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\ProfileUpdateRequest;
+use App\Models\User;
+use Exception;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Redirect;
+use Illuminate\Support\Facades\Validator;
 use Illuminate\View\View;
 
 class ProfileController extends Controller
@@ -35,25 +38,75 @@ class ProfileController extends Controller
         $request->user()->save();
         return Redirect::route('profile.edit')->with('status', 'profile-updated');
     }
+    /**
+     * Update the user's profile image.
+     */
+    public function updateImage(Request $request): RedirectResponse
+    {
+        // dd($request->request);
+        try {
+            $validated =  Validator::make($request->all(), [
+                'image' => 'file|mimes:jpg,jpeg,png,gif'
+
+
+            ]);
+            if (!$validated->stopOnFirstFailure()->fails()) {
+                $user = User::findOrFail(auth()->user()->id);
+                if ($request->image != null) {
+                    $originalName = $request->image->getClientOriginalName();
+                    $savedImage = time() . '_' . $originalName;
+                    $user->image = $savedImage;
+                    $user->update([
+                        'image' => $savedImage
+                    ]);
+
+                    $request->image->move(public_path('images/profile/users/'), $savedImage);
+                    $request->user()->save();
+                }
+            }
+
+
+            // toastr()->success('Student has been updated successfully!');
+            return Redirect::route('profile.edit')->with('status', 'profile image updated');
+        } catch (Exception $e) {
+            echo $e->getMessage();
+        }
+
+
+        return Redirect::route('profile.edit')->with('status', 'profile-updated');
+    }
 
     /**
      * Delete the user's account.
      */
-    public function destroy(Request $request): RedirectResponse
+    public function destroy(Request $request)
     {
-        $request->validateWithBag('userDeletion', [
-            'password' => ['required', 'current-password'],
-        ]);
 
-        $user = $request->user();
+        try {
+            // Form validation
 
-        Auth::logout();
+            $validated =  Validator::make($request->all(), [
+                'password' => ['required', 'current-password'],
 
-        $user->delete();
+            ]);
+            if ($validated->stopOnFirstFailure()->fails()) {
+                return response()->json([
+                    'error' => "password is not correct"
+                ]);
+            }
+            $user = $request->user();
+            Auth::logout();
+            $user->delete();
+            $request->session()->invalidate();
+            $request->session()->regenerateToken();
+            return response()->json([
+                'success' => 'email sent succssfully'
+            ]);
+        } catch (Exception $e) {
+            return  response()->json([
+                'error' => $e->getMessage()
 
-        $request->session()->invalidate();
-        $request->session()->regenerateToken();
-
-        return Redirect::to('/');
+            ]);
+        }
     }
 }
